Needed fuse.conf group readable by fuse for fs:/snapshots
New fuse3 seems to fix this by just having entire file world
readable.  Old fuse wanted it only fuse readable.  Looks like we're
fine with having it world readable - means we can do away with the
test that we needed a definition of the fuse group which didn't exist
if fuse package not installed...