check backups: use df -a to get all mountpoints
df without -a ignores "duplicate" mounts - where fsid is identical.
But in our LXC containers, we bind mount a whole bunch of directories
into our container that we want to backup.  Only the first was being
backed up by our LXC, because the rest are all on the same parent
filesystem.  But with -a, now we have to supply a whole bunch of extra
-x flags.