set up dnsmasq as a caching nameserver on all linux non pve hosts
attempted to set up systemd-resolved but didn't like the
NetworkManager setup and the fact that systemd-resolved is a bit of an
immature hack.  dnsmasq appears to do the job well with just as much
NetworkManager re-configuration, and no fragile host-dependent munging
of /etc/nsswitch.conf.  You can override by setting:
 - host_is_using_systemd_resolver: true
 - host_is_not_using_dnsmasq_cache: false