emacs new lockfile behaviour still annoys me
I just want emacs to detect when files have changed and ask me to
reload like it used to.  The current lockfile behaviour is just buggy:
edit a file locally and save it, and then try to open it remotely and
it asks whether I want to steal a lock.  No, I just to be able to have
emacs reliably detect that a file has changed underneath it.  This
*used* to work well.