want to install smartctlnvme into files/tree, rather than the sub repo
Since we install symlinks in files/tree to the remote host, we want to
resolve those symlinks when they point to locations that aren't going
to be copied to the resultant machine.  But we want to keep our sub
repos bundled.  So compile there, linking into here, rather than
linking from here into those repos