fridge1: PWMFrequency delegated to autoexec.be berry script
Could not convince PWMFrequency to apply my 4 settings independently
when done as a config - perhaps they're simply not saved in settings.
Nor when done as a backlog boot rule, perhaps because PWM is
initialised after switches are.  Didn't want to delay until after time
was initialised, because we want this to continue working when the
network is down.  Moved the initialisation to the berry script.