use negative lookahead regexp to munge /etc/default/zfs
Using it, I found a minor problem with the same scheme I used for
/etc/default/rkhunter - was replacing

foo="bar"

with:

foo="baaz"

which was fine with that file because those clauses were empty to
begin with.  Should make this more general, but first step would be to
allow specififcation of whether to use single or double quotes.