fix bash-preexec missed being invoked in subshells like ( 1 ; 2 )
Not invoking in subshells means that that the entire line doesn't end
up in your .bash_custom_history file ( at least until the next line is
invoked, in which case it catches up - if there is a next line ).

Fixed by testing that we're not apparently affected by bash-preexec#25
anymore, and setting __bp_enable_subshells=true before sourcing
bash-preexec

To guard against future bugs like this - perhaps each shell exit
should invoke the `history -a` line for safety before exiting.  This
was sort of already in place except didn't quite cater for this
scenario.  Untested because don't have the bug handy anymore :)