all hosts can now set modprobe_blacklist
Wanted to be able to blacklist modules from dirac, so extended scheme
that had been used in pve.  There was also a couple of bugs with the
comment field: first, we were using the wrong variable, so it was
never set, and two, when it did become set, we were putting a literal
"\\n" in there because of the way >- Block style with block chomping
indicator can't be told to insert a literal newline.

We're also bitten by the lazy evaluation non-implementation where
you'd end up trying to insert "__omit_place_holder__...." as a string.
Have to work around it by comparing against some dummy string such as
"not defined"