backport procps unstable for #1091673
We also don't pin those custom package versions for which there's no
alternative, or for which we allow newer packages to someday install.
In those cases, we'll have one play cause an upgrade, and this play
causes a downgrade, but that triggers us to remove this version from
our play because the upstream bug is then fixed.