longstanding typo in my /etc/lvm/lvm.conf fixed
Disk name exclusion was attempting to resolve to
/dev/.../ata-WUHWUH...

Also added wildcard to be sure to pick up on partitions.  Looking at
this because since October, lvscan in host is now picking up LVs from
hosts.  Change looks to be a proxmox or debian kernel change, but
haven't found the cause of it yet.

However, not having this fixed might have been responsible for backup
disks not having spun down recently.