Patch backuppc to not keep /var/lib/backuppc/LOG open permanently
Because I want to be able to automount /var/lib/backuppc to allow its
disks to spin down when not backing up, I need to not keep the one LOG
file open that it keeps open.  So as to not be too intrusive into the
backuppc code and rewrite every LOG caller, I just fork off a loop
that open-appends/writes/closes the file on demand.

(also backupppc machine needs to know how to turn on the backup backup
server, so remember to keep that file in sync with our repo now that
we've removed it from CVS control)