Remember to create the root directory itself when copying dirs
community.general.filetree is stupid and can't be convinced to return
the root directory itself.  We have to hack around it since we want it
in the right order (first, when creating), and we would rather have
access to all the stat fields without translating stat() ourselves
again.  Consequently, we don't bother with permissions on the root
directory, which is FIXME: nasty