create openwrt_upgrade.yml playbook to do openwrt firmware updates
Because of the way gekmihesg.openwrt works with its monkeypatching,
and we want the upgrade process and its reboots to be serialised so we
never lose all hosts at once, it's not simply a matter of calling the
upgrade process then the maintenance playbook or its roles, so there's
duplicated code between openwrt_upgrade.yml and
openwrt_maintenace.yml.  It's not elegant, but it seems to work.

Also removed gather_facts from openwrt_maintenance.yml since
discovered gather_facts was set and doesn't seem to be needed