fix history writing/reading/deduplication when last entry multiline
My infinite history with the optimisation to only read the last
$HISTSIZE entries was blowing up when multiline caused the first entry
to not be a datestamp, confusing readline's parsing of the file and
breaking up subsequent multiline input.  Deduplication was also
screwed because was just looking at the occurence of identical
strings, of which there might be many in a bash multiline scriptlet.