openwrt: change some IP addresses, and add ext.rather.puzzling.org
I wish to "delegate back" DNS for some fqdns back to upstream, and
this seems to be one way of doing it. I want to retrieve
ext.rather.puzzling.org from upstream now that internal resolution of
rather.puzzling.org maps back to our internal server.  It's not
pretty, and it encodes upstream DNS in several locations, but it's
better than the prior situation of having to run a daemon to
continually keep telling internal DNS what our external DNS address
is.