allow configuration of usb /overlay on openwrt
This playbook change allows once-off install of dependencies for an
USB disk for /overlay.  Still needs automation to configure the mount
rather than doing it manually whenever the dependency install is found
to be needed, which indicates a fresh upgrade has just been done.
Then we need to reinitiate the sync of the existing /overlay