39765c15 broke update of files on linux
Some files were getting reverted to root, predictably, when we just
wanted to keep their file ownership on linux.  So we're going to need
to back out *part* of the openwrt changes and code those permissions
explicitly in openwrt once we identify just what we were trying to do
there.