prepare for organisational cleanup of desktop packages
The list of packages to install on desktops is very large and
cumbersome, and if there's an installation error, it's difficult to
debug.  Let's break it up into a loop of related packages.  We might
be able to ignore the failure of installation of one group of packages
for instance, while carrying on with graceful installation of most of
the rest of the requirements.

This is the first preparation step to simplify the patches - just move
the array of packages to an array which gets looped over, of the same
array.