fridge: PID loop sped up in fridge
Because the response time of the fridge is quite quick (the cold plate has an effect on
the air directly, and contents only indirectly.  The contents might have a long response
time, but the air doesn't), reduce the I & D params of the PID loop by 75%.  Doesn't result in overshooting.