mygnuclient: merge current session's XAUTHORITY with server's
Since we isolated each terminal to its own XAUTHORITY (which was
implemented as a way of keeping XAUTHORITY from continuously growing
across a distributed NFS+SSH environment), we need to merge the
current ssh session's XAUTHORITY into that of the emacs server's.
Here we don't care about proliferation, because everything is in /tmp
and will be wiped soon enough.